backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0  # For async SQLite (dev)
asyncpg==0.29.0    # For async PostgreSQL (prod)

# Authentication
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# Email verification
fastapi-mail==1.4.1

# CORS
# (built into FastAPI)

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

python-dateutil~=2.9.0.post0
